+++ src/models/index.ts
@@ -1,7 +1,11 @@
-export { default as Chat } from "./chat.model";
-export { default as Dislikes } from "./dislikes.model";
-export { default as friendSearchProjection } from "./friendSearchProjection";
-export { default as Likes } from "./likes.model";
-export { default as Match } from "./match.model";
-export { default as Profile } from "./profile.model";
-export { Location, Preferences, ProfileDocument } from "./profile.model";
+export { default as Chat } from "../modules/chat/chat.model";
+export { default as Dislikes } from "../modules/deslikes/dislikes.model";
+export { default as friendSearchProjection } from "../modules/profile/friendSearchProjection";
+export { default as Likes } from "../modules/likes/likes.model";
+export { default as Match } from "../modules/match/match.model";
+export { default as Profile } from "../modules/profile/profile.model";
+export {
+  Location,
+  Preferences,
+  ProfileDocument,
+} from "../modules/profile/profile.model";
