+++ src/modules/chat/chat.model.ts
