+++ src/modules/profile/friendSearchProjection.ts
